remove useless template from showXPub modal
